requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
python-dotenv
requests
groq
psycopg2-binary
sqlalchemy
python-dotenv
